add test get kangle compile options so can disable some test case

+++ src/KWinCgiEnv.cpp
@@ -22,9 +22,9 @@ KWinCgiEnv::~KWinCgiEnv(void) {
 }
 bool KWinCgiEnv::add_env(const char* attr, size_t attr_len, const char* val, size_t val_len)
 {
-	s->write_all(attr, attr_len);
+	s->write_all(attr, (int)attr_len);
 	s->write_all(_KS("="));
-	s->write_all(val, val_len);
+	s->write_all(val, (int)val_len);
 	s->write_all(_KS("\0"));
 	return true;
 }
